requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
sentence-transformers
sqlalchemy
openai
requests
pyspellchecker